requirements.txt
torch==2.3.1
tensorrt==10.0.1
transformers==4.41.2
onnx==1.16.1
pillow==10.3.0
